srcs/requirements/matchmaking/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
channels==4.1.0
daphne==3.0.1
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3